Changes the amount of cards displayed per page
The CardContainer component now expects a prop called cardPageCount
that tells it how many cards to render per page. The DeckBuilder page
still only renders 9 cards at a time but the Card page now renders 27.

In the future, I would like to conditionally change the amount per page
based on the size of the screen. That way mobile compatability is more
reasonable. Either that or I will apply an overflow scroll to it at
certain resolutions.