Begins implementation of hamburger menu component logic
Adds a state called isMobile to the App component. The default is false,
and is changed to true when the resolution is adjust to be below 481px.

Currently it does not render as mobile if the resolution is 480px or
lower to start with and the hamburger menu component has not been
instantiated yet.

However, the navbar does change to the text "Hello" when it is adjusted
to mobile scale resolutions.