Installs font-awesome package and updates HamburgerMenu component
After installing the font-awesome package I added a triple bar icon to
the HamburgerMenu component. It now renders this icon on lower screen
resolutions. Need to tweak the styling for the icon as well as particular
ranges of resolutions still.

HamburgerMenu component is also rendered now instead of the previous
'Hello' string. Soon I will add an event listener to drop down the links
to my other "pages".