Changes addCard method
The addCard method previously used `selectedCards.concat(card)` which
is now replaced by `[...selectedCards, card]` because it seems more
proper.
I also thought that this was the fix to the DeckBuilder component
not updating but I think it has more to do with my understanding of how
functional components rerender. Will implement a fix hopefully later
after some sleep.