allCards method still not working
I believe the reason the allCards method is not working is that I am
not using a useEffect.
I can also just move the state for selectedCards
to the DeckBuilder component and manage that state there. Then, whenever
the selectedCards state changes, the useEffect dependency can be that
component and therefore will update it properly. I will test this
tomorrow.