Adds a useless useEffect hook to the DeckBuilder component
Currently this hook does nothing but it is the beginning of me deciding
where my state should reside. I am highly considering moving the
selectedCards state to the Deckbuilder component as a solution to the
updating problem I am having.
Also, only the Deckbuilder component needs
access to the selectedCards state and thus I believe it should belong to
this component and not the App.