Fixes Deckbuilder component and makes slight changes to Card component
The issue I faced in my deckbuilder component had nothing to do with
how I wrote and inherited the hooks and methods. I simply forgot to
pass the second CardContainer a cardsPerPage prop. Thus no cards would
display.

Card component changes are just the removal of the handleClick method
and the implementation of a function call inside of the onClick property.