Moves more specific Deckbuilder and Profile logic to their components
My main App component contained a lot of logic that could be stored in
the child components. As such, I moved the addCard, removeCard, and
saveDeck logic, as well as the selectedCards state, to my Deckbuilder
component.

Additionally, I moved the editDeck method inside of my Profile component
and changed how it works. Now, it changes the DeckId and redirects the
user to the Deckbuilder component.

Upon changing the deck_id state, the Deckbuilder component now fires off
a fetch request to get the specific deck's cards for populating the
selectedCards state.

Also, I have changed the fetch request URLs in my Sign Up and Login
components to properly hit my live API instead of the localhost.