Creates hooks folder and http hook
I want to create a reusable http request hook so I can remove a lot of
the rewritten code from my components. Right now the hook is configured
only for GET requests but once I confirm that it is working properly I
will implement a way to specify which kind of request to send.