Starts implementing custom hook into code
My custom useHttp hook is intended to replace all fetch calls in my
other components. As such, it currently takes a url and dependency list
and imports useState and useEffect to fetch data from a url, store the
data in state, and return that data every time the provided dependencies
change.

Still needs a bit of tweaking for optimal operation and future
requests will be accounted for.