Removes resetDeck function and polishes off http hook
I removed the resetDeck function because it was causing my hook to fire
off every time I went to a different page after loading a deck from the
user profile. Since the user, decks, and deck_id states are all managed
inside of the App component, where my hook is called, the data is fetched
every time one of those states updates. I'll be checking on what I can
do to change that. If anything.